Add a fix for Macs to translate horizontal scrolls to vertical
This change/fix is from the hscroll mod.
That mod replaces the entire mouse scroll handler with its custom
version, which only does this small change to the code anyway,
but by replacing the entire handler, it breaks for example all
the mouse scroll hooks and behavior in all malilib-based mods.