Snap Aim: Fix/improve behavior when Aim Lock is enabled with Snap Aim on
Also Snap Aim Lock doesn't snap to a stupid angle if outside the snap
region anymore, instead it will behave like an Aim Lock but only for
the axis (yaw or pitch) which is selected as the current Snap Aim mode.

Also moved some of the util code to a separate `SnapAimUtils` class.

# Conflicts:
#	src/main/java/fi/dy/masa/tweakeroo/mixin/MixinEntity.java
#	src/main/java/fi/dy/masa/tweakeroo/renderer/RenderUtils.java
#	src/main/java/fi/dy/masa/tweakeroo/util/MiscUtils.java